+++ app/models/colyseus-models/count.ts
@@ -26,4 +26,6 @@ export default class Count extends Schema implements ICount {
   @type("uint8") futureSightCount = 0
   @type("uint8") upgradeCount = 0
   @type("uint8") soulDewCount = 0
+  @type("uint8") healOrderCount = 0
+  @type("uint8") attackOrderCount = 0
 }