+++ app/public/news.md
@@ -1,7 +1,15 @@
-# Patch 3.5.1
+# Patch 3.6
 
 ## Newcomers
 
 Welcome to Phione ! Phione will evolve into a manaphy if you give him an aqua egg.
 
 ![phione](https://raw.githubusercontent.com/PMDCollab/SpriteCollab/master/portrait/0489/Normal.png) + ![aqua-egg](https://raw.githubusercontent.com/keldaanInteractive/pokemonAutoChess/master/app/public/dist/client/assets/item/AQUA_EGG.png) = ![manaphy](https://raw.githubusercontent.com/PMDCollab/SpriteCollab/master/portrait/0490/Normal.png)
+
+Welcome to Shucke !
+
+![shuckle](https://raw.githubusercontent.com/PMDCollab/SpriteCollab/master/portrait/0213/Normal.png)
+
+Welcome to Combee and Vespiqueen !
+
+![combee](https://raw.githubusercontent.com/PMDCollab/SpriteCollab/master/portrait/0415/Normal.png) ![vespiqueen](https://raw.githubusercontent.com/PMDCollab/SpriteCollab/master/portrait/0416/Normal.png)